fix: fix filter by cell not working for config related changes
