+++ src/pages/config/ConfigChangesPage.tsx
@@ -27,31 +27,6 @@ export function ConfigChangesPage() {
 
   const page = params.get("page") ?? "1";
   const pageSize = params.get("pageSize") ?? "200";
-  const configId = params.get("id") ?? undefined;
-  const changeSummary = params.get("summary") ?? undefined;
-  const source = params.get("source") ?? undefined;
-  const createdBy = params.get("created_by") ?? undefined;
-  const externalCreatedBy = params.get("external_created_by") ?? undefined;
-
-  const arbitraryFilter = useMemo(() => {
-    const filter = new Map<string, string>();
-    if (configId) {
-      filter.set("id", configId);
-    }
-    if (changeSummary) {
-      filter.set("summary", changeSummary);
-    }
-    if (source) {
-      filter.set("source", source);
-    }
-    if (createdBy) {
-      filter.set("created_by", createdBy);
-    }
-    if (externalCreatedBy) {
-      filter.set("external_created_by", externalCreatedBy);
-    }
-    return Object.fromEntries(filter);
-  }, [changeSummary, configId, createdBy, externalCreatedBy, source]);
 
   const { data, isLoading, error, isRefetching, refetch } =
     useGetAllConfigsChangesQuery({
@@ -130,10 +105,7 @@ export function ConfigChangesPage() {
             <InfoMessage message={errorMessage} />
           ) : (
             <>
-              <ConfigChangeFilters
-                paramsToReset={["page"]}
-                arbitraryFilters={arbitraryFilter}
-              />
+              <ConfigChangeFilters paramsToReset={["page"]} />
               <ConfigChangeTable
                 data={changes}
                 isLoading={isLoading}
